requirements.txt
Flask
requests
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pinecone>=3.0.0